move f2s to Wstring.cpp, fix arduino do not support int64 which was used by json c, first fix of getstarted

+++ AZ3166/src/cores/arduino/json-c/json_object.c
@@ -599,7 +599,7 @@ struct json_object* json_object_new_int(int32_t i)
 
 int32_t json_object_get_int(const struct json_object *jso)
 {
-  int64_t cint64;
+  int32_t cint64;
   enum json_type o_type;
 
   if(!jso) return 0;
