+++ AZ3166/src/cores/arduino/json-c/json_object_private.h
@@ -29,8 +29,8 @@ struct json_object
   struct printbuf *_pb;
   union data {
     json_bool c_boolean;
-    double c_double;
-    int64_t c_int64;
+    float c_double;
+    int32_t c_int64;
     struct lh_table *c_object;
     struct array_list *c_array;
     struct {
